+++ gvmtools/pyshell.py
@@ -16,34 +16,23 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import argparse
 import code
-import configparser
 import logging
 import os
 import sys
 
 from gvm import get_version as get_gvm_version
-from gvm.connections import (SSHConnection,
-                             TLSConnection,
-                             UnixSocketConnection,
-                             DEFAULT_UNIX_SOCKET_PATH,
-                             DEFAULT_TIMEOUT,
-                             DEFAULT_GVM_PORT)
 from gvm.protocols.latest import Gmp, Osp
 from gvm.transforms import EtreeCheckCommandTransform
 
 from gvmtools import get_version
 from gvmtools.helper import authenticate
+from gvmtools.parser import create_parser, create_connection, PROTOCOL_OSP
 
 __version__ = get_version()
 __api_version__ = get_gvm_version()
 
 logger = logging.getLogger(__name__)
-
-PROTOCOL_OSP = 'OSP'
-PROTOCOL_GMP = 'GMP'
-DEFAULT_PROTOCOL = PROTOCOL_GMP
 
 HELP_TEXT = """
     Command line tool to access services via GMP (Greenbone Management
@@ -103,152 +92,29 @@ class Arguments:
 
 
 def main():
-    parser = argparse.ArgumentParser(
-        prog='gvm-pyshell',
-        description=HELP_TEXT,
-        formatter_class=argparse.RawTextHelpFormatter,
-        add_help=False)
+    parser = create_parser(
+        description=HELP_TEXT, logfilename='gvm-pyshell.log')
 
-    subparsers = parser.add_subparsers(metavar='[connection_type]')
-    subparsers.required = True
-    subparsers.dest = 'connection_type'
+    parser.add_protocol_argument()
 
     parser.add_argument(
-        '-h', '--help', action='help',
-        help='Show this help message and exit')
-
-    parent_parser = argparse.ArgumentParser(add_help=False)
-
-    parent_parser.add_argument(
-        '-c', '--config', nargs='?', const='~/.config/gvm-tools.conf',
-        help='Configuration file path (default: ~/.config/gvm-tools.conf)')
-    args_before, remaining_args = parent_parser.parse_known_args()
-
-    defaults = {
-        'gmp_username': '',
-        'gmp_password': ''
-    }
-
-    # Retrieve data from config file
-    if args_before.config:
-        try:
-            config = configparser.SafeConfigParser()
-            path = os.path.expanduser(args_before.config)
-            config.read(path)
-            defaults = dict(config.items('Auth'))
-        except Exception as e: # pylint: disable=broad-except
-            print(str(e), file=sys.stderr)
-
-    parent_parser.set_defaults(**defaults)
-
-    parent_parser.add_argument(
-        '--timeout', required=False, default=DEFAULT_TIMEOUT, type=int,
-        help='Response timeout in seconds, or -1 to wait '
-             'indefinitely (default: %(default)s)')
-    parent_parser.add_argument(
-        '--log', nargs='?', dest='loglevel', const='INFO',
-        choices=['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'],
-        help='Activate logging (default level: %(default)s)')
-    parent_parser.add_argument(
         '-i', '--interactive', action='store_true', default=False,
         help='Start an interactive Python shell')
-    parent_parser.add_argument(
-        '--protocol', required=False, default=DEFAULT_PROTOCOL,
-        choices=[PROTOCOL_GMP, PROTOCOL_OSP],
-        help='Service protocol to use (default: %(default)s)')
-    parent_parser.add_argument('--gmp-username', help='Username for GMP service')
-    parent_parser.add_argument('--gmp-password', help='Password for GMP service')
-    parent_parser.add_argument(
+
+    parser.add_argument(
         'scriptname', nargs='?', metavar="SCRIPT",
         help='Path to script to be preloaded (example: myscript.gmp)')
-    parent_parser.add_argument(
+    parser.add_argument(
         'scriptargs', nargs='*', metavar="ARG",
         help='Arguments for preloaded script')
 
-    parser_ssh = subparsers.add_parser(
-        'ssh', help='Use SSH to connect to service',
-        parents=[parent_parser])
-    parser_ssh.add_argument('--hostname', required=True,
-                            help='Hostname or IP address')
-    parser_ssh.add_argument('--port', required=False,
-                            default=22, help='SSH port (default: %(default)s)')
-    parser_ssh.add_argument('--ssh-user', default='gmp',
-                            help='SSH username (default: %(default)s)')
+    args = parser.parse_args()
 
-    parser_tls = subparsers.add_parser(
-        'tls', help='Use TLS secured connection to connect to service',
-        parents=[parent_parser])
-    parser_tls.add_argument('--hostname', required=True,
-                            help='Hostname or IP address')
-    parser_tls.add_argument('--port', required=False,
-                            default=DEFAULT_GVM_PORT,
-                            help='GMP/OSP port (default: %(default)s)')
-    parser_tls.add_argument('--certfile', required=False, default=None,
-                            help='Path to the certificate file for client authentication')
-    parser_tls.add_argument('--keyfile', required=False, default=None,
-                            help='Path to key file for client authentication')
-    parser_tls.add_argument('--cafile', required=False, default=None,
-                            help='Path to CA certificate for server authentication')
-    parser_tls.add_argument('--no-credentials', required=False, default=False,
-                            help='Use only certificates for authentication')
+    if 'socket' in args.connection_type and args.sockpath:
+        print('The --sockpath parameter has been deprecated. Please use '
+              '--socketpath instead', file=sys.stderr)
 
-    parser_socket = subparsers.add_parser(
-        'socket', help='Use UNIX Domain socket to connect to service',
-        parents=[parent_parser])
-    parser_socket.add_argument(
-        '--sockpath', nargs='?', default=None,
-        help='Deprecated, use --socketpath instead')
-    parser_socket.add_argument(
-        '--socketpath', nargs='?', default=DEFAULT_UNIX_SOCKET_PATH,
-        help='Path to UNIX Domain socket (default: %(default)s)')
-
-    parser.add_argument(
-        '-V', '--version', action='version',
-        version='%(prog)s {version} (API version {apiversion})'.format(
-            version=__version__, apiversion=__api_version__),
-        help='Show version information and exit')
-
-    args = parser.parse_args(remaining_args)
-
-    # Sets the logging
-    if args.loglevel is not None:
-        level = logging.getLevelName(args.loglevel)
-        logging.basicConfig(filename='gvm-pyshell.log', level=level)
-
-    # If timeout value is -1, then the socket has no timeout for this session
-    if args.timeout == -1:
-        args.timeout = None
-
-    # Open the right connection. SSH at last for default
-    if 'socket' in args.connection_type:
-        socketpath = args.sockpath
-        if socketpath is None:
-            socketpath = args.socketpath
-        else:
-            print('The --sockpath parameter has been deprecated. Please use '
-                  '--socketpath instead', file=sys.stderr)
-
-        connection = UnixSocketConnection(
-            timeout=args.timeout,
-            path=socketpath
-        )
-    elif 'tls' in args.connection_type:
-        connection = TLSConnection(
-            timeout=args.timeout,
-            hostname=args.hostname,
-            port=args.port,
-            certfile=args.certfile,
-            keyfile=args.keyfile,
-            cafile=args.cafile,
-        )
-    else:
-        connection = SSHConnection(
-            timeout=args.timeout,
-            hostname=args.hostname,
-            port=args.port,
-            username=args.ssh_user,
-            password=''
-        )
+    connection = create_connection(**vars(args))
 
     transform = EtreeCheckCommandTransform()