+++ CHANGELOG.md
@@ -9,12 +9,19 @@ $ cd gvm-tools && git log
 
 # gvm-tools 2.0.0.beta2 (unreleased)
 
+# Configfile
+
+- The structure for the config file (default is ~/.config/gvm-tools.conf) has
+  changed. It's possible to set defaults for nearly all command line arguments.
+
 ## Other
 
 - The commandline help for `gvm-cli` and `gvm-pyshell` has been updated and
   made more consistent.
 - Fix a bug which caused `gvm-pyshell` to immediately re-enter interactive mode
   upon exiting it for the first time.
+- Renamed --ssh-user switch to --ssh-username
+- Added --ssh-password switch for ssh connection
 
 # gvm-tools 2.0.0.beta1 (13.11.2018)